Prevent form from clearing after an error. Move types to table file.
