+++ src/pages/staff/index.tsx
@@ -1,18 +1,14 @@
 import { trpc } from "@/client/lib/trpc";
 import React from "react";
-import PersonTable from "@/components/table/table";
-import { Para, ParaHeadCell } from "@/types/global";
+import PersonTable, { Para, ParaHeadCell } from "@/components/table/table";
 
 const Staff = () => {
   const utils = trpc.useContext();
   const { data: paras, isLoading } = trpc.case_manager.getMyParas.useQuery();
 
   const createPara = trpc.para.createPara.useMutation({
-    onSuccess: () => utils.case_manager.getMyParas.invalidate(),
-    onSettled: (data, error) => {
-      if (error) console.log(error.message);
-
-      assignParaToCaseManager.mutate({
+    onSuccess: async (data) => {
+      await assignParaToCaseManager.mutateAsync({
         para_id: data?.user_id as string,
       });
     },
@@ -20,20 +16,23 @@ const Staff = () => {
 
   const assignParaToCaseManager = trpc.case_manager.addPara.useMutation({
     onSuccess: () => utils.case_manager.getMyParas.invalidate(),
-    onError: (error) => console.log(error.message),
   });
 
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
     const data = new FormData(event.currentTarget);
 
-    createPara.mutate({
-      first_name: data.get("first_name") as string,
-      last_name: data.get("last_name") as string,
-      email: data.get("email") as string,
-    });
-    // resetting the form this way is only necessary if the form remains visible upon adding a person. due to Materials UI, the reset form(s) will show as "touched" (TT).
-    (event.target as HTMLFormElement).reset();
+    try {
+      await createPara.mutateAsync({
+        first_name: data.get("first_name") as string,
+        last_name: data.get("last_name") as string,
+        email: data.get("email") as string,
+      });
+      // resetting the form this way is only necessary if the form remains visible upon adding a person. due to Materials UI, the reset form(s) will show as "touched" (TT).
+      (event.target as HTMLFormElement).reset();
+    } catch (err) {
+      console.error(err);
+    }
   };
 
   if (isLoading) {
