+++ src/pages/students/index.tsx
@@ -1,14 +1,16 @@
 import React from "react";
 import { trpc } from "@/client/lib/trpc";
-import PersonTable from "@/components/table/table";
-import { StudentWithIep, StudentWithIepHeadcell } from "@/types/global";
+import PersonTable, {
+  StudentWithIep,
+  StudentWithIepHeadcell,
+} from "@/components/table/table";
 
 const Students = () => {
   const utils = trpc.useContext();
   const { data: students, isLoading } =
     trpc.case_manager.getMyStudentsAndIepInfo.useQuery();
 
-  const { mutate } = trpc.case_manager.addStudent.useMutation({
+  const createStudent = trpc.case_manager.addStudent.useMutation({
     onSuccess: () => utils.case_manager.getMyStudentsAndIepInfo.invalidate(),
     // TODO(tessa): In a future PR, we could change this to notification instead of browser alert
     onError: () =>
@@ -17,19 +19,22 @@ const Students = () => {
       ),
   });
 
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
     const data = new FormData(event.currentTarget);
-    const grade = Number(data.get("grade"));
 
-    mutate({
-      first_name: data.get("first_name") as string,
-      last_name: data.get("last_name") as string,
-      email: data.get("email") as string,
-      grade,
-    });
-    // resetting the form this way is only necessary if the form remains visible upon adding a person. due to Materials UI, the reset form(s) will show as "touched" (TT).
-    (event.target as HTMLFormElement).reset();
+    try {
+      await createStudent.mutateAsync({
+        first_name: data.get("first_name") as string,
+        last_name: data.get("last_name") as string,
+        email: data.get("email") as string,
+        grade: Number(data.get("grade")),
+      });
+      // resetting the form this way is only necessary if the form remains visible upon adding a person. due to Materials UI, the reset form(s) will show as "touched" (TT).
+      (event.target as HTMLFormElement).reset();
+    } catch (err) {
+      console.error(err);
+    }
   };
 
   const headCells: StudentWithIepHeadcell[] = [
